+++ src/errors/src/bl.js
@@ -1,8 +1,13 @@
 import BaseError from './base';
 
-export default class BLError extends BaseError {
-  constructor(message = 'The Business Logic script did not complete.'
-    + ' See debug message for details.', debug, code, kinveyRequestId) {
-    super('BLError', message, debug, code, kinveyRequestId);
-  }
+function BLError(message, debug, code, kinveyRequestId) {
+  this.name = 'BLError';
+  this.message = message || 'The Business Logic script did not complete.';
+  this.debug = debug || undefined;
+  this.code = code || undefined;
+  this.kinveyRequestId = kinveyRequestId || undefined;
+  this.stack = (new Error()).stack;
 }
+BLError.prototype = Object.create(BaseError.prototype);
+BLError.prototype.constructor = BLError;
+export default BLError;