+++ src/errors/src/corsDisabled.js
@@ -1,7 +1,13 @@
 import BaseError from './base';
 
-export default class CORSDisabledError extends BaseError {
-  constructor(message = 'Cross Origin Support is disabled for this application.', debug, code, kinveyRequestId) {
-    super('CORSDisabledError', message, debug, code, kinveyRequestId);
-  }
+function CORSDisabledError(message, debug, code, kinveyRequestId) {
+  this.name = 'CORSDisabledError';
+  this.message = message || 'Cross Origin Support is disabled for this application.';
+  this.debug = debug || undefined;
+  this.code = code || undefined;
+  this.kinveyRequestId = kinveyRequestId || undefined;
+  this.stack = (new Error()).stack;
 }
+CORSDisabledError.prototype = Object.create(BaseError.prototype);
+CORSDisabledError.prototype.constructor = CORSDisabledError;
+export default CORSDisabledError;