+++ src/errors/src/invalidCredentials.js
@@ -1,8 +1,13 @@
 import BaseError from './base';
 
-export default class InvalidCredentialsError extends BaseError {
-  constructor(message = 'Invalid credentials.'
-    + ' Please retry your request with correct credentials.', debug, code, kinveyRequestId) {
-    super('InvalidCredentialsError', message, debug, code, kinveyRequestId);
-  }
+function InvalidCredentialsError(message, debug, code, kinveyRequestId) {
+  this.name = 'InvalidCredentialsError';
+  this.message = message || 'Invalid credentials. Please retry your request with correct credentials.';
+  this.debug = debug || undefined;
+  this.code = code || undefined;
+  this.kinveyRequestId = kinveyRequestId || undefined;
+  this.stack = (new Error()).stack;
 }
+InvalidCredentialsError.prototype = Object.create(BaseError.prototype);
+InvalidCredentialsError.prototype.constructor = InvalidCredentialsError;
+export default InvalidCredentialsError;