+++ src/errors/src/invalidIdentifier.js
@@ -1,8 +1,13 @@
 import BaseError from './base';
 
-export default class InvalidIdentifierError extends BaseError {
-  constructor(message = 'One of more identifier names in the'
-    + ' request has an invalid format.', debug, code, kinveyRequestId) {
-    super('InvalidIdentifierError', message, debug, code, kinveyRequestId);
-  }
+function InvalidIdentifierError(message, debug, code, kinveyRequestId) {
+  this.name = 'InvalidIdentifierError';
+  this.message = message || 'One of more identifier names in the request has an invalid format.';
+  this.debug = debug || undefined;
+  this.code = code || undefined;
+  this.kinveyRequestId = kinveyRequestId || undefined;
+  this.stack = (new Error()).stack;
 }
+InvalidIdentifierError.prototype = Object.create(BaseError.prototype);
+InvalidIdentifierError.prototype.constructor = InvalidIdentifierError;
+export default InvalidIdentifierError;