+++ src/errors/src/kinveyInternalErrorStop.js
@@ -1,8 +1,13 @@
 import BaseError from './base';
 
-export default class KinveyInternalErrorStop extends BaseError {
-  constructor(message = 'The Kinvey server encountered an unexpected error.'
-    + ' Please contact support@kinvey.com for assistance.', debug, code, kinveyRequestId) {
-    super('KinveyInternalErrorStop', message, debug, code, kinveyRequestId);
-  }
+function KinveyInternalErrorStop(message, debug, code, kinveyRequestId) {
+  this.name = 'KinveyInternalErrorStop';
+  this.message = message || 'The Kinvey server encountered an unexpected error. Please contact support@kinvey.com for assistance.';
+  this.debug = debug || undefined;
+  this.code = code || undefined;
+  this.kinveyRequestId = kinveyRequestId || undefined;
+  this.stack = (new Error()).stack;
 }
+KinveyInternalErrorStop.prototype = Object.create(BaseError.prototype);
+KinveyInternalErrorStop.prototype.constructor = KinveyInternalErrorStop;
+export default KinveyInternalErrorStop;