+++ src/errors/src/noResponse.js
@@ -1,7 +1,13 @@
 import BaseError from './base';
 
-export default class NoResponseError extends BaseError {
-  constructor(message = 'No response was provided.', debug, code, kinveyRequestId) {
-    super('NoResponseError', message, debug, code, kinveyRequestId);
-  }
+function NoResponseError(message, debug, code, kinveyRequestId) {
+  this.name = 'NoResponseError';
+  this.message = message || 'No response was provided.';
+  this.debug = debug || undefined;
+  this.code = code || undefined;
+  this.kinveyRequestId = kinveyRequestId || undefined;
+  this.stack = (new Error()).stack;
 }
+NoResponseError.prototype = Object.create(BaseError.prototype);
+NoResponseError.prototype.constructor = NoResponseError;
+export default NoResponseError;