+++ src/errors/src/query.js
@@ -1,7 +1,13 @@
 import BaseError from './base';
 
-export default class QueryError extends BaseError {
-  constructor(message = 'An error occurred with the query.', debug, code, kinveyRequestId) {
-    super('QueryError', message, debug, code, kinveyRequestId);
-  }
+function QueryError(message, debug, code, kinveyRequestId) {
+  this.name = 'QueryError';
+  this.message = message || 'An error occurred with the query.';
+  this.debug = debug || undefined;
+  this.code = code || undefined;
+  this.kinveyRequestId = kinveyRequestId || undefined;
+  this.stack = (new Error()).stack;
 }
+QueryError.prototype = Object.create(BaseError.prototype);
+QueryError.prototype.constructor = QueryError;
+export default QueryError;