+++ src/errors/src/staleRequest.js
@@ -1,7 +1,13 @@
 import BaseError from './base';
 
-export default class StaleRequestError extends BaseError {
-  constructor(message = 'The time window for this request has expired.', debug, code, kinveyRequestId) {
-    super('StaleRequestError', message, debug, code, kinveyRequestId);
-  }
+function StaleRequestError(message, debug, code, kinveyRequestId) {
+  this.name = 'StaleRequestError';
+  this.message = message || 'The time window for this request has expired.';
+  this.debug = debug || undefined;
+  this.code = code || undefined;
+  this.kinveyRequestId = kinveyRequestId || undefined;
+  this.stack = (new Error()).stack;
 }
+StaleRequestError.prototype = Object.create(BaseError.prototype);
+StaleRequestError.prototype.constructor = StaleRequestError;
+export default StaleRequestError;