+++ src/errors/src/sync.js
@@ -1,7 +1,13 @@
 import BaseError from './base';
 
-export default class SyncError extends BaseError {
-  constructor(message = 'An error occurred during sync.', debug, code, kinveyRequestId) {
-    super('SyncError', message, debug, code, kinveyRequestId);
-  }
+function SyncError(message, debug, code, kinveyRequestId) {
+  this.name = 'SyncError';
+  this.message = message || 'An error occurred during sync.';
+  this.debug = debug || undefined;
+  this.code = code || undefined;
+  this.kinveyRequestId = kinveyRequestId || undefined;
+  this.stack = (new Error()).stack;
 }
+SyncError.prototype = Object.create(BaseError.prototype);
+SyncError.prototype.constructor = SyncError;
+export default SyncError;