+++ src/request/src/cache.js
@@ -1,11 +1,10 @@
 import Promise from 'es6-promise';
-import UrlPattern from 'url-pattern';
+import Queue from 'promise-queue';
 import url from 'url';
-import localStorage from 'local-storage';
 import cloneDeep from 'lodash/cloneDeep';
 
 import Client from 'src/client';
-import { KinveyError, NotFoundError } from 'src/errors';
+import { KinveyError } from 'src/errors';
 import Query from 'src/query';
 import Aggregation from 'src/aggregation';
 import { isDefined, Log } from 'src/utils';
@@ -16,6 +15,9 @@ import { CacheRack } from './rack';
 const usersNamespace = process.env.KINVEY_USERS_NAMESPACE || 'user';
 const activeUserCollectionName = process.env.KINVEY_USER_ACTIVE_COLLECTION_NAME || 'kinvey_active_user';
 const activeUsers = {};
+
+Queue.configure(Promise);
+const queue = new Queue(1, Infinity);
 
 /**
  * @private
@@ -66,12 +68,12 @@ export default class CacheRequest extends Request {
 
   set url(urlString) {
     super.url = urlString;
-    const pathname = global.escape(url.parse(urlString).pathname);
-    const pattern = new UrlPattern('(/:namespace)(/)(:appKey)(/)(:collection)(/)(:entityId)(/)');
-    const { appKey, collection, entityId } = pattern.match(pathname) || {};
-    this.appKey = appKey;
-    this.collection = collection;
-    this.entityId = entityId;
+    const pathname = global.decodeURIComponent(url.parse(urlString).pathname);
+    const urlParts = pathname.replace(/^\//, '').split('/');
+    // "pathname" has the following form: "/namespace/appKey/collection/id"
+    this.appKey = urlParts[1];
+    this.collection = urlParts[2];
+    this.entityId = urlParts[3];
   }
 
   execute() {
@@ -116,46 +118,30 @@ export default class CacheRequest extends Request {
   }
 
   static loadActiveUser(client = Client.sharedInstance()) {
-    const request = new CacheRequest({
-      method: RequestMethod.GET,
-      url: url.format({
-        protocol: client.apiProtocol,
-        host: client.apiHost,
-        pathname: `/${usersNamespace}/${client.appKey}/${activeUserCollectionName}`
-      })
-    });
-    return request.execute()
-      .then(response => response.data)
-      .then((users) => {
-        if (users.length > 0) {
-          return users[0];
-        }
-
-        return null;
-      })
-      .then((activeUser) => {
-        // Try local storage
-        if (isDefined(activeUser) === false) {
-          return CacheRequest.loadActiveUserLegacy(client);
-        }
-
-        return activeUser;
-      })
-      .then((activeUser) => {
-        return CacheRequest.setActiveUser(client, activeUser);
-      })
-      .then((activeUser) => {
-        Log.debug('Load active user', client, activeUser);
-        activeUsers[client.appKey] = activeUser;
-        return activeUser;
+    return queue.add(() => {
+      const request = new CacheRequest({
+        method: RequestMethod.GET,
+        url: url.format({
+          protocol: client.apiProtocol,
+          host: client.apiHost,
+          pathname: `/${usersNamespace}/${client.appKey}/${activeUserCollectionName}`
+        })
       });
-  }
+      return request.execute()
+        .then(response => response.data)
+        .then((users) => {
+          if (users.length > 0) {
+            return users[0];
+          }
 
-  static loadActiveUserLegacy(client = Client.sharedInstance()) {
-    const activeUser = CacheRequest.getActiveUserLegacy(client);
-    Log.debug('Load active user legacy', client, activeUser);
-    activeUsers[client.appKey] = activeUser;
-    return activeUser;
+          return null;
+        })
+        .then((activeUser) => {
+          Log.debug('Load active user', client, activeUser);
+          activeUsers[client.appKey] = activeUser;
+          return activeUser;
+        });
+    });
   }
 
   static getActiveUser(client = Client.sharedInstance()) {
@@ -164,63 +150,18 @@ export default class CacheRequest extends Request {
     return activeUser;
   }
 
-  static getActiveUserLegacy(client = Client.sharedInstance()) {
-    try {
-      const activeUser = localStorage.get(`${client.appKey}kinvey_user`);
-      Log.debug('Get active user legacy', client, activeUser);
-      return activeUser;
-    } catch (error) {
-      Log.error('Get active user legacy', client, error);
-      return null;
-    }
-  }
-
-  static setActiveUser(client = Client.sharedInstance(), user) {
+  static setActiveUser(client = Client.sharedInstance(), activeUser) {
     Log.debug('Set active user', client, user);
-    let promise = Promise.resolve(null);
-    const activeUser = CacheRequest.getActiveUser(client);
+    return queue.add(() => {
+      let promise = Promise.resolve(null);
+      const prevActiveUser = CacheRequest.getActiveUser(client);
 
-    if (isDefined(activeUser)) {
-      // Delete from local storage (legacy)
-      CacheRequest.setActiveUserLegacy(client, null);
-
-      // Delete from memory
-      activeUsers[client.appKey] = null;
-
-      // Delete from cache
-      const request = new CacheRequest({
-        method: RequestMethod.DELETE,
-        url: url.format({
-          protocol: client.apiProtocol,
-          host: client.apiHost,
-          pathname: `/${usersNamespace}/${client.appKey}/${activeUserCollectionName}/${activeUser._id}`
-        })
-      });
-      promise = request.execute()
-        .then(response => response.data)
-        .catch((error) => {
-          if (error instanceof NotFoundError) {
-            return null;
-          }
-
-          throw error;
-        });
-    }
-
-    return promise
-      .then(() => {
-        if (isDefined(user) === false) {
-          return null;
-        }
-
-        // Remove sensitive data from user
-        delete user.password;
+      if (isDefined(activeUser)) {
+        // Remove sensitive data from activeUser
+        delete activeUser.password;
 
         // Save to memory
-        activeUsers[client.appKey] = user;
-
-        // Save to local storage (legacy)
-        CacheRequest.setActiveUserLegacy(client, user);
+        activeUsers[client.appKey] = activeUser;
 
         // Save to cache
         const request = new CacheRequest({
@@ -230,27 +171,34 @@ export default class CacheRequest extends Request {
             host: client.apiHost,
             pathname: `/${usersNamespace}/${client.appKey}/${activeUserCollectionName}`
           }),
-          body: user
+          body: activeUser
         });
-        return request.execute()
+        promise = request.execute()
           .then(response => response.data);
-      })
-      .then(() => user);
-  }
-
-  static setActiveUserLegacy(client = Client.sharedInstance(), user) {
-    try {
-      localStorage.remove(`${client.appKey}kinvey_user`);
-
-      if (isDefined(user)) {
-        localStorage.set(`${client.appKey}kinvey_user`, user);
+      } else {
+        // Delete from memory
+        delete activeUsers[client.appKey];
       }
 
-      Log.debug('Set active user legacy', client, user);
-      return true;
-    } catch (error) {
-      Log.error('Set active user legacy', client, error);
-      return false;
-    }
+      return promise
+        .then(() => {
+          if (isDefined(prevActiveUser) && (isDefined(activeUser) === false || prevActiveUser._id !== activeUser._id)) {
+            // Delete from cache
+            const request = new CacheRequest({
+              method: RequestMethod.DELETE,
+              url: url.format({
+                protocol: client.apiProtocol,
+                host: client.apiHost,
+                pathname: `/${usersNamespace}/${client.appKey}/${activeUserCollectionName}/${prevActiveUser._id}`
+              })
+            });
+            return request.execute()
+              .catch(() => activeUser);
+          }
+
+          return activeUser;
+        })
+        .then(() => activeUser);
+    });
   }
 }